+++ bin/main.ml
@@ -12,13 +12,13 @@ module Input_format = struct
   let all = [ JSONSchema; OpenAPI ]
 end
 
-let generate_atd input_format paths =
+let generate_atd state input_format paths =
   let generate =
     match input_format with
-    | Input_format.JSONSchema -> Generator.make_atd_of_jsonschema
-    | OpenAPI -> Generator.make_atd_of_openapi
+    | Input_format.JSONSchema -> Generator.make_atd_of_jsonschema ~state
+    | OpenAPI -> Generator.make_atd_of_openapi ~state
   in
-  print_endline (Generator.base (String.concat " " (List.map Filename.basename paths)));
+  print_endline (Generator.base state (String.concat " " (List.map Filename.basename paths)));
   let root =
     match paths with
     | [ _ ] -> `Default
@@ -44,8 +44,42 @@ let input_format_term =
 
 let main =
   let doc = "Generate an ATD file from a list of JSON Schema / OpenAPI document" in
+  let state_term =
+    Term.(
+      const (fun skip_doc pad toplevel_types avoid_dangling_refs json_ocaml_type ->
+        Generator.
+          {
+            with_doc = not skip_doc;
+            protect_against_duplicates = (if pad then Some (ref []) else None);
+            toplevel_types;
+            avoid_dangling_refs;
+            json_ocaml_type;
+          }
+      )
+      $ Arg.(value (flag (info [ "skip-doc" ] ~doc:"Skip documentation annotations.")))
+      $ Arg.(value (flag (info [ "protect-against-duplicates" ] ~doc:"Make sure no duplicate types are generated.")))
+      $ (const (function
+           | [] -> `All
+           | some -> `Only some
+           )
+        $ Arg.(
+            value (opt_all string [] (info [ "only-matching" ] ~docv:"REGEXP" ~doc:"Only output types matching REGEXP."))
+          )
+        )
+      $ Arg.(value (flag (info [ "avoid-dangling-refs" ] ~doc:"Convert dangling refs to json.")))
+      $ Arg.(
+          let keyword = enum [ "module", `Module; "from", `From ] in
+          value
+            (opt (t3 ~sep:':' keyword string string) Generator.default_state.json_ocaml_type
+               (info [ "json-ocaml-type" ] ~docv:"KEYWORD:MODULE.PATH:TYPE-NAME"
+                  ~doc:"Use an alternate Mod.type for `json`, e.g. from:My_mod.Submod:json_type."
+               )
+            )
+        )
+    )
+  in
   let paths = Arg.(non_empty & pos_all file [] & info [] ~docv:"FILES" ~doc) in
-  let term = Term.(const generate_atd $ input_format_term $ paths) in
+  let term = Term.(const generate_atd $ state_term $ input_format_term $ paths) in
   let info = Cmd.info "jsonschema2atd" ~doc ~version:(Version.get ()) in
   Cmd.v info term